requirements.txt
flask==3.0.2
flask-restx==1.3.0
python-dotenv==1.0.1
pyyaml==6.0.1
openapi-python-client==0.19.0
httpx==0.27.0
attrs==25.3.0 
